+++ flyway-dialect/src/main/java/tech/ydb/flywaydb/database/YdbSchema.java
@@ -19,7 +19,7 @@ public class YdbSchema extends Schema<YdbDatabase, YdbTable> {
 
     @Override
     protected boolean doExists() {
-        return false;
+        return true; // Dummy schema exists
     }
 
     @Override
@@ -29,15 +29,16 @@ public class YdbSchema extends Schema<YdbDatabase, YdbTable> {
 
     @Override
     protected void doCreate() {
+        // Do nothing, YDB doesn't support schemas
     }
 
     @Override
     protected void doDrop() {
+        // Do nothing, YDB doesn't support schemas
     }
 
     @Override
     protected void doClean() {
-
     }
 
     @Override