+++ liquibase-dialect/src/main/java/tech/ydb/liquibase/type/BoolTypeYdb.java
@@ -1,6 +1,5 @@
 package tech.ydb.liquibase.type;
 
-import liquibase.change.core.LoadDataChange;
 import liquibase.datatype.DataTypeInfo;
 import liquibase.datatype.LiquibaseDataType;
 
@@ -18,14 +17,4 @@ import liquibase.datatype.LiquibaseDataType;
         priority = LiquibaseDataType.PRIORITY_DATABASE
 )
 public class BoolTypeYdb extends BaseTypeYdb {
-
-    @Override
-    public LoadDataChange.LOAD_DATA_TYPE getLoadTypeName() {
-        return LoadDataChange.LOAD_DATA_TYPE.BOOLEAN;
-    }
-
-    @Override
-    protected String objectToSql(Object value) {
-        return value.toString();
-    }
 }