+++ liquibase-dialect/src/main/java/tech/ydb/liquibase/type/IntTypeYdb.java
@@ -1,6 +1,5 @@
 package tech.ydb.liquibase.type;
 
-import liquibase.change.core.LoadDataChange;
 import liquibase.datatype.DataTypeInfo;
 import liquibase.datatype.LiquibaseDataType;
 
@@ -9,20 +8,10 @@ import liquibase.datatype.LiquibaseDataType;
  */
 @DataTypeInfo(
         name = "Int32",
-        aliases = {"int", "integer", "java.sql.Types.INTEGER", "java.lang.Integer", "int4", "int32"},
+        aliases = {"int", "integer", "java.sql.Types.INTEGER", "java.lang.Integer", "int4"},
         minParameters = 0,
         maxParameters = 0,
         priority = LiquibaseDataType.PRIORITY_DATABASE
 )
 public class IntTypeYdb extends BaseTypeYdb {
-
-    @Override
-    public LoadDataChange.LOAD_DATA_TYPE getLoadTypeName() {
-        return LoadDataChange.LOAD_DATA_TYPE.NUMERIC;
-    }
-
-    @Override
-    public String objectToSql(Object value) {
-        return value.toString();
-    }
 }