+++ liquibase-dialect/src/main/java/tech/ydb/liquibase/type/TimeTypeYdb.java
@@ -1,8 +1,10 @@
 package tech.ydb.liquibase.type;
 
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
+import liquibase.database.Database;
 import liquibase.datatype.DataTypeInfo;
 import liquibase.datatype.LiquibaseDataType;
 
@@ -11,10 +13,7 @@ import liquibase.datatype.LiquibaseDataType;
  */
 @DataTypeInfo(
         name = "Datetime",
-        aliases = {
-                "java.util.Date", "time", "java.sql.Types.TIME", "java.sql.Time",
-                "timetz", "java.sql.Types.TIME_WITH_TIMEZONE"
-        },
+        aliases = {"time", "java.sql.Types.TIME", "java.sql.Time"},
         minParameters = 0,
         maxParameters = 0,
         priority = LiquibaseDataType.PRIORITY_DATABASE
@@ -22,9 +21,23 @@ import liquibase.datatype.LiquibaseDataType;
 public class TimeTypeYdb extends BaseTypeYdb {
 
     @Override
-    protected String objectToSql(Object value) {
+    public String objectToSql(Object value, Database database) {
+        if ((value == null) || "null".equalsIgnoreCase(value.toString())) {
+            return "NULL";
+        }
+
+        if (value instanceof Timestamp) {
+            return database.getDateTimeLiteral((Timestamp) value);
+        }
+
         return "DATETIME('" + LocalDateTime.parse(value.toString())
                 .atZone(ZoneId.of("UTC"))
                 .format(DateTimeFormatter.ISO_INSTANT) + "')";
+    }
+
+    // not used
+    @Override
+    protected String objectToSql(Object value) {
+        return "";
     }
 }