+++ liquibase-dialect/src/test/java/tech/ydb/liquibase/YdbDatabaseLiquibaseAllTypesTableTest.java
@@ -25,11 +25,11 @@ public class YdbDatabaseLiquibaseAllTypesTableTest extends BaseTest {
                         "tinyint_column INT8, " +
                         "float_column FLOAT, " +
                         "double_column DOUBLE, " +
-                        "decimal_column DECIMAL(22, 9), " +
+                        "decimal_column DECIMAL(22,9), " +
                         "uint8_column UINT8, " +
                         "uint16_column UINT16, " +
-                        "unit32_column UINT32, " +
-                        "unit64_column UINT64, " +
+                        "uint32_column UINT32, " +
+                        "uint64_column UINT64, " +
                         "text_column TEXT, " +
                         "binary_column BYTES, " +
                         "json_column JSON, " +