+++ liquibase-dialect/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.0.2 ##
+
+* Added duration format for `Interval` type and ISO for time types when loading data into tables.
+
+## 1.0.1 ##
+
+* Scan select databasechangelog
+
+## 1.0.0 ##
+
+* Fixed bug with distributed lock (setAutoCommit = false)
+* Supported insert changeset with all YDB types
+
+---
+
 ## 0.9.7 ##
 
 * Supported NOT NULL statement